// frontend/apps/web/src/domains/api/generated/model/jobRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * PostPilot
 * OpenAPI spec version: 0.1.0
 */
import type { JobReadSourceUrl } from './jobReadSourceUrl';
import type { JobSourceKind } from './jobSourceKind';
import type { JobStatus } from './jobStatus';
import type { JobReadErrorMessage } from './jobReadErrorMessage';
import type { JobReadTitle } from './jobReadTitle';
import type { JobReadCompany } from './jobReadCompany';
import type { JobReadLocation } from './jobReadLocation';
import type { RemotePolicy } from './remotePolicy';
import type { Seniority } from './seniority';
import type { JobReadSalaryRange } from './jobReadSalaryRange';
import type { JobReadSummary } from './jobReadSummary';
import type { JobRequirementRead } from './jobRequirementRead';

export interface JobRead {
  id: string;
  source_url: JobReadSourceUrl;
  source_kind: JobSourceKind;
  raw_text: string;
  status: JobStatus;
  error_message: JobReadErrorMessage;
  title: JobReadTitle;
  company: JobReadCompany;
  location: JobReadLocation;
  remote_policy: RemotePolicy;
  seniority: Seniority;
  tech_stack: string[];
  salary_range: JobReadSalaryRange;
  summary: JobReadSummary;
  requirements: JobRequirementRead[];
  created_at: string;
  updated_at: string;
}