// frontend/apps/web/src/domains/api/generated/model/remotePolicy.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * PostPilot
 * OpenAPI spec version: 0.1.0
 */

export type RemotePolicy = typeof RemotePolicy[keyof typeof RemotePolicy];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RemotePolicy = {
  onsite: 'onsite',
  hybrid: 'hybrid',
  remote: 'remote',
  unspecified: 'unspecified',
} as const;