// frontend/apps/web/src/domains/api/generated/model/seniority.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * PostPilot
 * OpenAPI spec version: 0.1.0
 */

export type Seniority = typeof Seniority[keyof typeof Seniority];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Seniority = {
  junior: 'junior',
  mid: 'mid',
  senior: 'senior',
  lead: 'lead',
  principal: 'principal',
  unspecified: 'unspecified',
} as const;